#189 merge functionparameter and parameter files
