--- docs/source/expression_tree/function_parameter.rst
@@ -1,5 +1,0 @@
-Function Parameter
-==================
-
-.. autoclass:: pybamm.FunctionParameter
-  :members:
