+++ docs/source/expression_tree/index.rst
@@ -5,7 +5,6 @@ Expression Tree
 
   symbol
   parameter
-  function_parameter
   variable
   independent_variable
   scalar
