+++ docs/source/expression_tree/parameter.rst
@@ -4,3 +4,5 @@ Parameter
 .. autoclass:: pybamm.Parameter
   :members:
 
+.. autoclass:: pybamm.FunctionParameter
+  :members:
