+++ pybamm/__init__.py
@@ -67,7 +67,6 @@ from .expression_tree.concatenations import (
 )
 from .expression_tree.array import Array
 from .expression_tree.matrix import Matrix
-from .expression_tree.parameter import Parameter
 from .expression_tree.unary_operators import (
     UnaryOperator,
     Negate,
@@ -82,8 +81,8 @@ from .expression_tree.unary_operators import (
     div,
     surf,
 )
+from .expression_tree.parameter import Parameter, FunctionParameter
 from .expression_tree.broadcasts import Broadcast, NumpyBroadcast
-from .expression_tree.function_parameter import FunctionParameter
 from .expression_tree.scalar import Scalar
 from .expression_tree.variable import Variable
 from .expression_tree.independent_variable import (
