+++ tests/test_expression_tree/test_parameter.py
@@ -17,6 +17,21 @@ class TestParameter(unittest.TestCase):
         self.assertEqual(c.domain[0], "test")
 
 
+class TestFunctionParameter(unittest.TestCase):
+    def test_function_parameter_init(self):
+        var = pybamm.Variable("var")
+        func = pybamm.FunctionParameter("func", var)
+        self.assertEqual(func.name, "func")
+        self.assertEqual(func.children[0].id, var.id)
+        self.assertEqual(func.domain, [])
+        self.assertEqual(func.diff_variable, None)
+
+    def test_function_parameter_diff(self):
+        var = pybamm.Variable("var")
+        func = pybamm.FunctionParameter("a", var).diff(var)
+        self.assertEqual(func.diff_variable, var)
+
+
 if __name__ == "__main__":
     print("Add -v for more debug output")
     import sys
